moved Svg element to core
